fixed compilier errors

+++ Assets/Code/Scripts/LevelDesign/Hazards/Spikes.cs
@@ -3,7 +3,7 @@ using UnityEngine.SceneManagement;
 
 public class Spikes : MonoBehaviour
 {
-    [SerializeField] private float damage = 5000;
+    [SerializeField] private int damage = 5000;
     private const string FIRST_LEVEL_NAME = "L1_Redo_Nika";
     private void OnTriggerEnter2D(Collider2D collision)
     {
